move SSOButton to components folder
