+++ src/App.svelte
@@ -10,7 +10,7 @@
     import { filterEvent } from './eventFilter';
     import type { QueryCstNode } from './generated/chevrotain_dts';
     import Message from './Message.svelte';
-    import SSOButton from './SSOButton.svelte';
+    import SSOButton from './components/SSOButton.svelte';
     import LogoutButton from './LogoutButton.svelte';
     import SSORevokeButton from './SSORevokeButton.svelte';
     import { validate } from './session';
